requirements.txt
PyYAML==6.0
sh==1.14.2
termcolor==1.1.0
schedule